Update example site, storybook and docs
