+++ packages/search-box/src/SearchBoxLabel.tsx
@@ -1,5 +1,6 @@
 import { Text } from '@ag.ds-next/text';
 import { visuallyHiddenStyles } from '@ag.ds-next/a11y';
+import { mapSpacing } from '@ag.ds-next/core';
 
 export type SearchBoxLabelprops = {
 	children: string;
@@ -15,8 +16,9 @@ export const SearchBoxLabel = ({
 	return (
 		<Text
 			as="label"
+			fontWeight="bold"
 			htmlFor={htmlFor}
-			css={visible ? undefined : visuallyHiddenStyles}
+			css={visible ? { marginBottom: mapSpacing(0.5) } : visuallyHiddenStyles}
 		>
 			{children}
 		</Text>
